hackathon-todo-phase4/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg2-binary>=2.9.9
pydantic>=2.10.0
pydantic[email]>=2.10.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
httpx>=0.28.1
passlib[bcrypt]>=1.7.4
sqlmodel>=0.0.22
google-generativeai>=0.8.0
openai>=1.0.0
